% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtins.R
\name{read_pcap}
\alias{read_pcap}
\title{Read a PCAP file}
\usage{
read_pcap(fname, filter, layers = 3)
}
\arguments{
\item{fname}{path to the capture file.}

\item{filter}{BPF filter to apply for reading the capture.}

\item{layers}{number of layers to decode.}
}
\description{
Read a PCAP file
}
\examples{
## Not run:
fname <- system.file("pcaps/http.cap", package="Rtins")
pcap <- read_pcap(fname)
head(pcap)
summary(pcap)

## End(Not run)
}
