% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtins.R
\name{sniff_pcap}
\alias{sniff_pcap}
\title{Sniff a PCAP file}
\usage{
sniff_pcap(iface, filter, num, layers = 3)
}
\arguments{
\item{iface}{network interface to sniff.}

\item{filter}{BPF filter to apply for reading the capture.}

\item{num}{number of packets to collect}

\item{layers}{number of layers to decode.}
}
\description{
Sniff a PCAP file
}
\examples{
## Not run:
interface<-'eth0'
pcap <- sniff_pcap(interface)
head(pcap)
summary(pcap)
head(sniff_pcap(interface, filter = "udp and dst port 53", num = 20))

## End(Not run)
}
